+++ FlOpEDT/TTapp/TTConstraints/core_constraints.py
@@ -31,8 +31,7 @@ from base.timing import TimeInterval
 from base.models import CourseStartTimeConstraint, Department, TimeGeneralSettings, TransversalGroup, Week
 from django.db import models
 
-from TTapp.TTConstraint import TTConstraint
-from TTapp.ilp_constraints.constraint import Constraint
+from TTapp.TTConstraints.TTConstraint import TTConstraint
 from TTapp.ilp_constraints.constraint_type import ConstraintType
 from TTapp.ilp_constraints.constraints.instructorConstraint import InstructorConstraint
 from TTapp.ilp_constraints.constraints.slotInstructorConstraint import SlotInstructorConstraint
@@ -198,7 +197,7 @@ class NoSimultaneousGroupCourses(TTConstraint):
                             
         return jsondict
 
-    def enrich_model(self, ttmodel, week, ponderation=1):
+    def enrich_ttmodel(self, ttmodel, week, ponderation=1):
         relevant_slots = slots_filter(ttmodel.wdb.availability_slots, week=week)
         relevant_basic_groups = considered_basic_groups(self, ttmodel)
         # Count the number of transversal groups
@@ -271,7 +270,7 @@ class ScheduleAllCourses(TTConstraint):
     tutors = models.ManyToManyField('people.Tutor', blank=True)
     course_types = models.ManyToManyField('base.CourseType', blank=True)
                                 
-    def enrich_model(self, ttmodel, week, ponderation=1):
+    def enrich_ttmodel(self, ttmodel, week, ponderation=1):
         relevant_basic_groups = considered_basic_groups(self, ttmodel)
         considered_courses = set(c for bg in relevant_basic_groups
                                  for c in ttmodel.wdb.all_courses_for_basic_group[bg])
@@ -321,7 +320,7 @@ class AssignAllCourses(TTConstraint):
     course_types = models.ManyToManyField('base.CourseType', blank=True)
     pre_assigned_only = models.BooleanField(default=False, verbose_name=_('Pre-assigned courses only'))
 
-    def enrich_model(self, ttmodel, week, ponderation=1):
+    def enrich_ttmodel(self, ttmodel, week, ponderation=1):
         relevant_basic_groups = considered_basic_groups(self, ttmodel)
         considered_courses = set(c for bg in relevant_basic_groups
                                  for c in ttmodel.wdb.all_courses_for_basic_group[bg])
@@ -497,7 +496,7 @@ class ConsiderTutorsUnavailability(TTConstraint):
                             jsondict["status"] = _("KO")
         return jsondict
 
-    def enrich_model(self, ttmodel, week, ponderation=1):
+    def enrich_ttmodel(self, ttmodel, week, ponderation=1):
         considered_tutors = set(ttmodel.wdb.instructors)
         if self.tutors.exists():
             considered_tutors &= set(self.tutors.all())