+++ FlOpEDT/people/models.py
@@ -30,6 +30,7 @@ from base.models import Department
 from base.timing import Day
 from django.core.validators import MinValueValidator, MaxValueValidator
 
+
 # Create your models here.
 
 
@@ -55,33 +56,33 @@ class User(AbstractUser):
     def has_department_perm(self, department, admin=False):
         """
         Does the user have access to a specific department
-        
-        admin=True    Check if the user can access to the 
+
+        admin=True    Check if the user can access to the
                       department admin
         """
         if self.is_superuser:
             return True
-        
-        return (self.is_tutor 
+
+        return (self.is_tutor
                 and department in self.departments.all()
                 and (not admin
                      or
-                     UserDepartmentSettings.objects\
+                     UserDepartmentSettings.objects \
                      .get(user=self,
-                          department=department)\
+                          department=department) \
                      .is_admin
-                    )
-               )
+                     )
+                )
 
     def has_perm(self, perm, obj=None):
         "Does the user have a specific permission?"
         # Simplest possible answer: Yes, always
-        return  self.is_staff
+        return self.is_staff
 
     def has_module_perms(self, app_label):
         "Does the user have permissions to view the app `app_label`?"
         # Simplest possible answer: Yes, always
-        return  self.is_staff
+        return self.is_staff
 
     def uni_extended(self):
         ret = self.username + '<'
@@ -93,15 +94,20 @@ class User(AbstractUser):
         ret += '(' + str(self.rights) + ')'
         return ret
 
+    ###
+    #   #Allows to get the user's preferred theme in the "base.html" file
+    ###
+    @property
+    def get_theme(self):
+        return self.themes_preference.theme
+
     class Meta:
         ordering = ['username', ]
-
-    
 
 
 class UserDepartmentSettings(models.Model):
     """
-    This model allows to add additionnal settings to the 
+    This model allows to add additionnal settings to the
     relation between User and Department
     """
     user = models.ForeignKey(User, on_delete=models.CASCADE)
@@ -125,16 +131,10 @@ class Tutor(User):
                               choices=TUTOR_CHOICES,
                               verbose_name="Status",
                               default=FULL_STAFF)
-    pref_hours_per_day = models.PositiveSmallIntegerField(
-        verbose_name="How many hours per day would you prefer ?",
-        default=4)
-    max_hours_per_day = models.PositiveSmallIntegerField(
-        verbose_name="How many hours per day can you suffer ?",
-        default=9)
 
     def uni_extended(self):
         ret = super(Tutor, self).uni_extended()
-        ret += '-' + self.status + '-' + 'P' + str(self.pref_hours_per_day) + 'M' + str(self.max_hours_per_day)
+        ret += '-' + self.status
         return ret
 
     class Meta:
@@ -187,6 +187,26 @@ class BIATOS(Tutor):
     class Meta:
         verbose_name = 'BIATOS'
 
+
+class TutorPreference(models.Model):
+    tutor = models.OneToOneField('Tutor',
+                                 on_delete=models.CASCADE,
+                                 related_name='preferences')
+    pref_hours_per_day = models.PositiveSmallIntegerField(
+        verbose_name="How many hours per day would you prefer ?",
+        default=4)
+    max_hours_per_day = models.PositiveSmallIntegerField(
+        verbose_name="How many hours per day can you suffer ?",
+        default=9)
+    min_hours_per_day = models.PositiveSmallIntegerField(
+        verbose_name="Under how many hours would you prefer to avoid to have class?",
+        default=0)
+
+    def __str__(self):
+        ret = f"{self.tutor} - P{self.pref_hours_per_day} - M{self.pref_hours_per_day} - m{self.min_hours_per_day}"
+        return ret
+
+
 # --- Notes sur Prof ---
 #    MinDemiJournees=models.BooleanField(
 #       verbose_name="Min Demi-journées?", default=False)
@@ -199,7 +219,7 @@ class BIATOS(Tutor):
 
 class Student(User):  # for now: representative
     belong_to = models.ManyToManyField('base.GenericGroup',
-                                               blank=True)
+                                       blank=True)
 
     def __str__(self):
         return str(self.username)
@@ -225,13 +245,13 @@ class Preferences(models.Model):
         return float(self.morning_weight)
 
     def get_evening_weight(self):
-        return float(1-self.morning_weight)
+        return float(1 - self.morning_weight)
 
     def get_free_half_day_weight(self):
         return float(self.free_half_day_weight)
 
     def get_light_day_weight(self):
-        return float(1-self.free_half_day_weight)
+        return float(1 - self.free_half_day_weight)
 
     class Meta:
         abstract = True
@@ -271,10 +291,10 @@ class GroupPreferences(Preferences):
                 local_hole_weight += student_pref.hole_weight
                 local_eat_weight += student_pref.eat_weight
             # To calculate the average of each attributs
-            self.morning_weight = local_morning_weight/nb_student_prefs
-            self.free_half_day_weight = local_free_half_day_weight/nb_student_prefs
-            self.hole_weight = local_hole_weight/nb_student_prefs
-            self.eat_weight = local_eat_weight/nb_student_prefs
+            self.morning_weight = local_morning_weight / nb_student_prefs
+            self.free_half_day_weight = local_free_half_day_weight / nb_student_prefs
+            self.hole_weight = local_hole_weight / nb_student_prefs
+            self.eat_weight = local_eat_weight / nb_student_prefs
             self.save()
 
 
@@ -284,6 +304,14 @@ class NotificationsPreferences(models.Model):
                                 related_name='notifications_preference')
     nb_of_notified_weeks = models.PositiveSmallIntegerField(default=0)
 
+###
+# Save the user's preferred theme in the data base
+###
+class ThemesPreferences(models.Model):
+    user = models.OneToOneField('User',
+                                on_delete=models.CASCADE,
+                                related_name='themes_preference')
+    theme = models.CharField(max_length=50, default='White')
 
 class UserPreferredLinks(models.Model):
     user = models.OneToOneField('User',
@@ -294,7 +322,7 @@ class UserPreferredLinks(models.Model):
 
     def __str__(self):
         return self.user.username + ' : ' + \
-            ' ; '.join([str(l) for l in self.links.all()])
+               ' ; '.join([str(l) for l in self.links.all()])
 
 
 class PhysicalPresence(models.Model):