+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 ### Added
+- RoomPonderation objects to allow TTModel without room assignation
+- RoomModel to assign Rooms
+- Method reassign_rooms uses RoomModel
+- Possibility to pre-assign rooms and/or post-assign rooms in TTModel
+
+- Possibility to add a preferred theme in the use's preferences
 
 ### Changed